requirements.txt
Flask==3.0.3
matplotlib==3.8.3
numpy==2.1.1
pandas==2.2.3
Pillow==10.4.0
scikit_learn==1.4.2
